products table is populated with products.csv

+++ schema.sql
@@ -8,7 +8,7 @@ CREATE TABLE products (
   item_id INTEGER(11) NOT NULL AUTO_INCREMENT,
   product_name VARCHAR(50) NOT NULL,
   department_name VARCHAR(50) NOT NULL,
-  price INTEGER(10) NOT NULL,
+  price DECIMAL(13,2) NOT NULL,
   stock_quantity VARCHAR(50),
   PRIMARY KEY (item_id)
 );
